% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_elem}
\alias{add_elem}
\title{Add elements to a vector}
\usage{
add_elem(n)
}
\arguments{
\item{n}{An integer giving the final size of the vector}
}
\value{
A vector of size \code{n}
}
\description{
Add elements to a vector
}
