% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bubble_sort}
\alias{bubble_sort}
\title{Naive implementation of bubble sort algorithm}
\usage{
bubble_sort(x)
}
\arguments{
\item{x}{A numeric vector which values are sorted in increasing order}
}
\value{
A numeric vector with values sorted in increasing order
}
\description{
Naive implementation of bubble sort algorithm
}
