% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_poly_area}
\alias{calc_poly_area}
\title{Return the area of an irregular polygon}
\usage{
calc_poly_area(x, y)
}
\arguments{
\item{x}{A vector of x-coordinates}

\item{y}{A vector of y-coordinates (positive \code{y}-vaues mean upwards)}
}
\value{
The area of the polygon
}
\description{
Return the area of an irregular polygon
}

