% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_tree_struct}
\alias{create_tree_struct}
\title{Illustrates definition and usage of a simple class with only data members}
\usage{
create_tree_struct(dbh, height, no_stems, spec_name)
}
\arguments{
\item{dbh}{The trees diameter at breast height (double)}

\item{height}{The trees height (double)}

\item{no_stems}{The number of stems the tree has (int)}

\item{spec_name}{The species of the tree (string)}
}
\value{
No return values, simply printing to the R-console
}
\description{
Illustrates definition and usage of a simple class with only data members
}

