% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dem_cond}
\alias{dem_cond}
\title{Conditional operator}
\usage{
dem_cond(x)
}
\arguments{
\item{x}{An int}
}
\value{
Nothing to return, just printing to the console
}
\description{
Conditional operator
}

