% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dem_if_else}
\alias{dem_if_else}
\title{An if-else example}
\usage{
dem_if_else(dbh)
}
\arguments{
\item{dbh}{Diameter at breast height in cm}
}
\value{
A diameter class
}
\description{
An if-else example
}

