% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fact_cpp}
\alias{fact_cpp}
\title{Calculate the factorial of an integer number}
\usage{
fact_cpp(n)
}
\arguments{
\item{n}{An integer number}
}
\value{
The factorial of \code{n}
}
\description{
Meant to use only for teaching purposes not for replacing the built-in
function \code{factorial()}.
}
\details{
A C++ implementation using a \code{for} loop.
}
\examples{
   fact_cpp(10);
}
