% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_example.R
\name{fact_r}
\alias{fact_r}
\title{Calculate the factorial of an integer number}
\usage{
fact_r(n)
}
\arguments{
\item{n}{An integer number}
}
\value{
The factorial of \code{n}
}
\description{
Meant to use only for teaching purposes not for replacing the built-in
function \code{factorial()}.
}
\details{
Internally a \code{for} loop is used.
}
\examples{
   fact_r(10);
}
