% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_functions.R
\name{fact_rec_r}
\alias{fact_rec_r}
\title{The factorial of an integer number}
\usage{
fact_rec_r(n)
}
\arguments{
\item{n}{An integer number}
}
\value{
The factorial of \code{n}
}
\description{
Meant to use only for teaching purposes not for replacing the built-in
function \code{factorial()}.
}
\details{
A base R implementation using recursive programming.
}
\examples{
fact_rec_r(10);
}
