% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_first_neg_dw}
\alias{find_first_neg_dw}
\title{Find first negative element in a vector and return its position}
\usage{
find_first_neg_dw(x)
}
\arguments{
\item{x}{A vector of double}
}
\value{
Index of the first negative value
}
\description{
Find first negative element in a vector and return its position
}
