% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_knn}
\alias{find_knn}
\title{Search nearest neighbours}
\usage{
find_knn(Y, X, M, k)
}
\arguments{
\item{X}{Matrix with auxiliary variables of the reference set. Rows are the
single observations and variables are stored in columns.}

\item{M}{A square matrix defining the distance measure. When \code{M} is
the identity matrix, the Euclidean distance results.}

\item{k}{The number of neighbours}
}
\description{
Search nearest neighbours
}
