% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_max}
\alias{find_max}
\title{Find the maximum from a list of values}
\usage{
find_max(x)
}
\arguments{
\item{x}{A vector holding a number of values}
}
\value{
The largest number in \code{x}
}
\description{
Find the maximum from a list of values
}
