% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gen_fib_cpp}
\alias{gen_fib_cpp}
\title{A function to calculate fibonacci numbers (implemented in C++)}
\usage{
gen_fib_cpp(n)
}
\arguments{
\item{n}{The length of the sequence}
}
\value{
A vector containing \code{n} fibonacci numbers
}
\description{
A function to calculate fibonacci numbers (implemented in C++)
}
\examples{
n <- 20;
gen_fib_cpp(n);
}
