% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_functions.R
\name{gen_fib_r}
\alias{gen_fib_r}
\title{A function to calculate fibonacci numbers}
\usage{
gen_fib_r(n)
}
\arguments{
\item{n}{The length of the sequence}
}
\value{
A vector containing \code{n} fibonacci numbers
}
\description{
A function to calculate fibonacci numbers
}
\examples{
n <- 20;
gen_fib_r(n);
}

