% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{illustrate_scope}
\alias{illustrate_scope}
\title{Illustrate scoping rules}
\usage{
illustrate_scope()
}
\value{
No return values, simply printing to the R-console
}
\description{
Illustrate scoping rules
}
