% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mix_signed_unsigned}
\alias{mix_signed_unsigned}
\title{Show effect of mixing signed and unsigned types}
\usage{
mix_signed_unsigned()
}
\value{
No return values, simply printing to the R-console
}
\description{
Show effect of mixing signed and unsigned types
}
