% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{round_cpp}
\alias{round_cpp}
\title{Demonstrate rounding in C++
R's built-in \code{round()} function shows a
different behaviour than the \code{round()} function of C++}
\usage{
round_cpp(x)
}
\arguments{
\item{x}{A double value to be rounded}
}
\value{
The rounded integer represented as double
}
\description{
Demonstrate rounding in C++
R's built-in \code{round()} function shows a
different behaviour than the \code{round()} function of C++
}

