% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{show_pointers}
\alias{show_pointers}
\title{Function to illustrate pointers}
\usage{
show_pointers()
}
\value{
No return values, simply printing to the R-console
}
\description{
Function to illustrate pointers
}
