% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{show_references}
\alias{show_references}
\title{Function to illustrate references}
\usage{
show_references()
}
\value{
No return values, simply printing to the R-console
}
\description{
Function to illustrate references
}

