% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{show_type_ranges}
\alias{show_type_ranges}
\title{Print C++ data type ranges to the R-console}
\usage{
show_type_ranges()
}
\value{
No return values, simply printing to the R-console
}
\description{
Print C++ data type ranges to the R-console
}

