% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{squ_elem}
\alias{squ_elem}
\title{Square vector elements}
\usage{
squ_elem(v)
}
\arguments{
\item{v}{A vector of doubles}
}
\value{
A double containing the sum of all elements of \code{v}
}
\description{
Square vector elements
}

