% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sum_elem}
\alias{sum_elem}
\title{Sums all elements in a vector}
\usage{
sum_elem(v)
}
\arguments{
\item{v}{A vector of doubles}
}
\value{
A double containing the sum of all elements of \code{v}
}
\description{
Sums all elements in a vector
}
