% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{to_upper_subscript}
\alias{to_upper_subscript}
\title{Convert all characters in a string to upper case}
\usage{
to_upper_subscript(str)
}
\arguments{
\item{str}{A string}
}
\value{
The same string as \code{str} but with upper case letters
}
\description{
Convert all characters in a string to upper case
}

