% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{type_conversions}
\alias{type_conversions}
\title{Print C++ data type conversions}
\usage{
type_conversions()
}
\value{
No return values, simply printing to the R-console
}
\description{
Print C++ data type conversions
}

