% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{use_list}
\alias{use_list}
\title{Demonstrate the use of a list object from the Rcpp package}
\usage{
use_list(x, y)
}
\arguments{
\item{x}{A numeric matrix}

\item{y}{A numeric vector}
}
\value{
A named list containing both \code{x} and \code{y}.
}
\description{
Demonstrate the use of a list object from the Rcpp package
}
