% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{use_matrix}
\alias{use_matrix}
\title{Example function for demonstrating Rcpp's NumericMatrix class}
\usage{
use_matrix(x)
}
\arguments{
\item{x}{A numeric matrix}
}
\value{
A numeric matrix created inside the function. A subset of \code{x}
  is printed on the console
}
\description{
Example function for demonstrating Rcpp's NumericMatrix class
}
