% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsintro_package.R
\docType{package}
\name{spsintro}
\alias{spsintro}
\alias{.onUnload}
\alias{spsintro-package}
\title{spsintro}
\usage{
.onUnload(libpath)
}
\description{
Code snippets and functions used for 'Introduction to Scientific Programming
and Simulation' course
}
