% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{use_list}
\alias{use_list}
\title{Example function to demonstrate Rcpp's list class}
\usage{
use_list(x, y)
}
\arguments{
\item{x}{A numeric matrix}

\item{y}{A numeric vector}
}
\value{
A list with \code{x} and \code{y} as its elements
}
\description{
Example function to demonstrate Rcpp's list class
}
