% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{use_matrix}
\alias{use_matrix}
\title{Demonstrate the use of a numeric matrix object from the Rcpp package}
\usage{
use_matrix(x)
}
\arguments{
\item{x}{A numeric matrix}
}
\value{
A new matrix with the same dimensions as \code{x} filled with zeroes
}
\description{
Demonstrate the use of a numeric matrix object from the Rcpp package
}
